requirements.txt
requests==2.12.3
tabulate==0.7.5
termcolor==1.1.0
docopt==0.6.2
pyfiglet==0.7.5
python-firebase==1.2
sqlalchemy==1.1.4